Move products and plans logic to a hook.
