+++ client/jetpack-cloud/sections/partner-portal/issue-license-v2/licenses-form/index.tsx
@@ -1,24 +1,13 @@
-import { getQueryArg } from '@wordpress/url';
 import { useTranslate } from 'i18n-calypso';
-import { useCallback, useEffect, useMemo, useContext } from 'react';
+import { useCallback, useContext } from 'react';
 import QueryProductsList from 'calypso/components/data/query-products-list';
-import {
-	isJetpackBundle,
-	isWooCommerceProduct,
-	isWpcomHostingProduct,
-} from 'calypso/jetpack-cloud/sections/partner-portal/lib';
 import LicenseBundleCard from 'calypso/jetpack-cloud/sections/partner-portal/license-bundle-card';
 import LicenseProductCard from 'calypso/jetpack-cloud/sections/partner-portal/license-product-card';
-import { useDispatch, useSelector } from 'calypso/state';
-import useProductsQuery from 'calypso/state/partner-portal/licenses/hooks/use-products-query';
-import { getAssignedPlanAndProductIDsForSite } from 'calypso/state/partner-portal/licenses/selectors';
-import {
-	addSelectedProductSlugs,
-	clearSelectedProductSlugs,
-} from 'calypso/state/partner-portal/products/actions';
+import { useSelector } from 'calypso/state';
 import { getDisabledProductSlugs } from 'calypso/state/partner-portal/products/selectors';
 import IssueLicenseContext from '../context';
 import useSubmitForm from '../hooks/use-submit-form';
+import useProductAndPlans from './hooks/use-product-and-plans';
 import LicensesFormSection from './sections';
 import type { AssignLicenceProps } from '../../types';
 import type {
@@ -34,79 +23,22 @@ export default function LicensesForm( {
 	quantity = 1,
 }: AssignLicenceProps ) {
 	const translate = useTranslate();
-	const dispatch = useDispatch();
-
-	const { data, isLoading: isLoadingProducts } = useProductsQuery();
 
 	const { selectedLicenses, setSelectedLicenses } = useContext( IssueLicenseContext );
 
-	let allProducts = data;
-	const addedPlanAndProducts = useSelector( ( state ) =>
-		selectedSite ? getAssignedPlanAndProductIDsForSite( state, selectedSite.ID ) : null
-	);
-
-	// Filter products & plan that are already assigned to a site
-	if ( selectedSite && addedPlanAndProducts && allProducts ) {
-		allProducts = allProducts.filter(
-			( product ) => ! addedPlanAndProducts.includes( product.product_id )
-		);
-	}
-
-	const bundles =
-		allProducts?.filter(
-			( { family_slug }: { family_slug: string } ) => family_slug === 'jetpack-packs'
-		) || [];
-	const backupAddons =
-		allProducts
-			?.filter(
-				( { family_slug }: { family_slug: string } ) => family_slug === 'jetpack-backup-storage'
-			)
-			.sort( ( a, b ) => a.product_id - b.product_id ) || [];
-	const products =
-		allProducts?.filter(
-			( { family_slug }: { family_slug: string } ) =>
-				family_slug !== 'jetpack-packs' &&
-				family_slug !== 'jetpack-backup-storage' &&
-				! isWooCommerceProduct( family_slug ) &&
-				! isWpcomHostingProduct( family_slug )
-		) || [];
-	const wooExtensions =
-		allProducts?.filter( ( { family_slug }: { family_slug: string } ) =>
-			isWooCommerceProduct( family_slug )
-		) || [];
-
-	// If the user comes from the flow for adding a new payment method during an attempt to issue a license
-	// after the payment method is added, we will make an attempt to issue the chosen license automatically.
-	const defaultProductSlugs = useMemo(
-		() => getQueryArg( window.location.href, 'products' )?.toString().split( ',' ),
-		[]
-	);
-
-	useEffect( () => {
-		// Select the slugs included in the URL
-		defaultProductSlugs &&
-			dispatch(
-				addSelectedProductSlugs(
-					// Filter the bundles and select only individual products
-					defaultProductSlugs.filter( ( slug ) => ! isJetpackBundle( slug ) )
-				)
-			);
-
-		// Clear all selected slugs when navigating away from the page to avoid persisting the data.
-		return () => {
-			dispatch( clearSelectedProductSlugs() );
-		};
-	}, [ dispatch, defaultProductSlugs ] );
+	const {
+		allProductsAndBundles,
+		isLoadingProducts,
+		bundles,
+		backupAddons,
+		products,
+		wooExtensions,
+		suggestedProductSlugs,
+	} = useProductAndPlans( { selectedSite } );
 
 	const disabledProductSlugs = useSelector< PartnerPortalStore, string[] >( ( state ) =>
-		getDisabledProductSlugs( state, allProducts ?? [] )
+		getDisabledProductSlugs( state, allProductsAndBundles ?? [] )
 	);
-
-	// We need the suggested products (i.e., the products chosen from the dashboard) to properly
-	// track if the user purchases a different set of products.
-	const suggestedProductSlugs = getQueryArg( window.location.href, 'product_slug' )
-		?.toString()
-		.split( ',' );
 
 	const handleSelectBundleLicense = useCallback(
 		( product: APIProductFamilyProduct ) => {
